Unify system health
