+++ database_manager/src/lib.rs
@@ -98,10 +98,9 @@ fn parse_client_config<E: EthSpec>(
     cli_args: &ArgMatches,
     _env: &Environment<E>,
 ) -> Result<ClientConfig, String> {
-    let mut client_config = ClientConfig {
-        data_dir: get_data_dir(cli_args),
-        ..Default::default()
-    };
+    let mut client_config = ClientConfig::default();
+
+    client_config.set_data_dir(get_data_dir(cli_args));
 
     if let Some(freezer_dir) = clap_utils::parse_optional(cli_args, "freezer-dir")? {
         client_config.freezer_db_path = Some(freezer_dir);
