+++ validator_client/Cargo.toml
@@ -60,4 +60,5 @@ reqwest = { version = "0.11.0", features = ["json","stream"] }
 url = "2.2.2"
 malloc_utils = { path = "../common/malloc_utils" }
 sysinfo = "0.26.5"
+system_health = { path = "../common/system_health" }
